ANDROID: kbuild: add support for compiling external device trees
Similar to external modules, the device tree should also be allowed to
be compiled using out-of-tree source. This would allow one to further
separate the core kernel from the vendor-specific device trees. To do
this, we just need to allow the user to set the dtstree and
DTC_INCLUDE build properties in the build environment.

To setup the external device tree, you can follow these steps:

1) Move the device tree to your out-of-tree project.
2) Move the dt-bindings to your out-of-tree project.
3) Configure your build environment to set the following variables:

  dtstree=/path/to/device/tree
  DTC_INCLUDE=/path/to/device/tree/dtc/include-prefixes

If you are using build.sh, then you can use DTS_EXT_DIR to define the
dtstree.

Bug: 210036798
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: Ibf05544ee8250a68b882328817a267518ae478b5
Signed-off-by: Lee Jones <joneslee@google.com>